group-collar-insert: Correct stretched funnel
Two locations used a z-stretched funnel, presumably to avoid the shared
plane shearing effect in preview mode. This resulted in distortion of
the inner wall and ramp walls as the diameter variables weren't adjusted
to compensate for the additional height.

Correct this by using the height corresponding to the measured
diameters, and add an element to the difference group to clear the
shared plane (sized to have no impact on the model itself). Update the
ramp bevel and ramp wall fading to account for the funnel changes.

Signed-off-by: Darren Hart <darren@dvhart.com>